% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_corr_fun.R
\name{absolute_corr_fun}
\alias{absolute_corr_fun}
\title{Check covariate balance}
\usage{
absolute_corr_fun(w, c, nthread)
}
\arguments{
\item{w}{A vector of observed continuous exposure variable.}

\item{c}{A data frame or matrix of observed covariates variable.}

\item{nthread}{Number of available threads to use.}
}
\value{
The function returns a list saved the measure related to covariate balance
\code{absolute_corr}: the absolute correlations for each pre-exposure
covairates;
\code{mean_absolute_corr}: the average absolute correlations for all
pre-exposure covairates.
}
\description{
Checks covariate balance based on absolute correlations for given data sets.
}
\keyword{internal}
