% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_weighted_corr_fun.R
\name{absolute_weighted_corr_fun}
\alias{absolute_weighted_corr_fun}
\title{Check covariate balance (weighted correlation)}
\usage{
absolute_weighted_corr_fun(w, ipw, c)
}
\arguments{
\item{w}{A vector of observed continuous exposure variable.}

\item{ipw}{A vector of weights.}

\item{c}{A data frame or matrix of observed covariates variable.}
}
\value{
The function returns a list saved the measure related to covariate balance
\code{absolute_corr}: the absolute correlations for each pre-exposure
covairates;
\code{mean_absolute_corr}: the average absolute correlations for all
pre-exposure covairates.
}
\description{
Checks covariate balance based on absolute weighted correlations for
given data sets.
}
\keyword{internal}
