% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_balance.R
\name{AbsoluteCorrFun}
\alias{AbsoluteCorrFun}
\title{Check covariate balance based on absolute correlations for given data sets.}
\usage{
AbsoluteCorrFun(w, c)
}
\arguments{
\item{w}{A vector of observed continuous exposure variable.}

\item{c}{A data frame or matrix of observed covariates variable.}
}
\value{
The function returns a list saved the measure related to covariate balance
\code{absolute_corr}: the absolute correlations for each pre-exposure
covairates;
\code{mean_absolute_corr}: the average absolute correlations for all
pre-exposure covairates.
}
\description{
Check covariate balance based on absolute correlations for given data sets.
}
