% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.gpsm_erf}
\alias{autoplot.gpsm_erf}
\title{A helper function for gpsm_erf object}
\usage{
\method{autoplot}{gpsm_erf}(object, ...)
}
\arguments{
\item{object}{A gpsm_erf object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot object.
}
\description{
A helper function to plot gpsm_erf object using ggplot2 package.
}
\keyword{internal}
