% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.gpsm_pspop}
\alias{autoplot.gpsm_pspop}
\title{A helper function for gpsm_pspop object}
\usage{
\method{autoplot}{gpsm_pspop}(object, ...)
}
\arguments{
\item{object}{A gpsm_pspop object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot object.
}
\description{
A helper function to plot gpsm_pspop object using ggplot2 package.
}
\keyword{internal}
