% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_args_compile_pseudo_pop}
\alias{check_args_compile_pseudo_pop}
\title{Check compile_pseudo_pop function arguments}
\usage{
check_args_compile_pseudo_pop(ci_appr, ...)
}
\arguments{
\item{...}{Additional arguments to successfully run the selected ci_appr.}

\item{data_obj}{A S3 object with \code{dataset} as required parameter.}

\item{ci.appr}{The selected causal inference approach.}
}
\value{
Returns True if passes all checks, successfully. Otherwise raises ERROR.
}
\description{
additional arguments are provided.
}
\keyword{internal}
