% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_args_use_cov_transformers}
\alias{check_args_use_cov_transformers}
\title{Check Covariate Balance Transformers Argument}
\usage{
check_args_use_cov_transformers(use_cov_transform, transformers)
}
\arguments{
\item{use_cov_transform}{A logical value (TRUE/FALSE) to use covariate balance
transforming.}

\item{transformers}{A list of transformers.}
}
\value{
TRUE if passes all tests.
}
\description{
Checks Covariate Balance Transformers in terms of using them and available
transformers.
}
\keyword{internal}
