% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_covar_balance.R
\name{check_covar_balance}
\alias{check_covar_balance}
\title{Check covariate balance}
\usage{
check_covar_balance(pseudo_pop, ci_appr, nthread = 1, optimized_compile, ...)
}
\arguments{
\item{pseudo_pop}{The generated pseudo population. In the following format:
\itemize{
\item 1st column: outcome (Y)
\item 2nd column: exposure (w)
\item 3rd column: gps
\item 4th column to the end: covariates (c)
}}

\item{ci_appr}{The causal inference approach.}

\item{nthread}{The number of available threads.}

\item{optimized_compile}{If TRUE, use optimized compile approach.}

\item{...}{Additional arguments passed to different models.}
}
\value{
Returns True if the pseudo population meet the covariate balance test
requirements.
}
\description{
Checks the covariate balance of
}
\keyword{internal}
