% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_kolmogorov_smirnov.R
\name{check_kolmogorov_smirnov}
\alias{check_kolmogorov_smirnov}
\title{Check Kolmogorov-Smirnov (KS) statistics}
\usage{
check_kolmogorov_smirnov(
  w,
  c,
  ci_appr,
  optimized_compile,
  counter_weight = NULL,
  nthread = 1
)
}
\arguments{
\item{w}{A vector of observed continuous exposure variable.}

\item{c}{A data.frame of observed covariates variable.}

\item{ci_appr}{The causal inference approach.}

\item{optimized_compile}{If TRUE, use optimized compile approach.}

\item{counter_weight}{A weight vector in different situations. If the
matching approach is selected, it is an integer data.table of counters.
In the case of the weighting approach, it is weight data.table.}

\item{nthread}{The number of available threads.}
}
\value{
output object is list including:
\itemize{
\item ks_stat
\item maximal_val
\item mean_val
\item median_val
}
}
\description{
Checks the Kolmogorov-Smirnov (KS) statistics for exposure and confounders in
the psuedo-population
}
\keyword{internal}
