% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{CheckArgs}
\alias{CheckArgs}
\title{Check Additional Arguments}
\usage{
CheckArgs(pred.model, ci.appr, running.appr, ...)
}
\arguments{
\item{pred.model}{The prediction model.}

\item{ci.appr}{The causal inference approach.}

\item{running.appr}{The running approach.}

\item{...}{Additional named arguments passed.}
}
\value{
TRUE if requirements are met. Raises error otherwise.
}
\description{
Checks additional arguments that user needs to provide for different
prediction models.
}
\keyword{internal}
