% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{CheckArgsCPseudoPop}
\alias{CheckArgsCPseudoPop}
\title{Check CompilePseudoPop function arguments}
\usage{
CheckArgsCPseudoPop(ci.appr, ...)
}
\arguments{
\item{ci.appr}{The selected causal inference approach.}

\item{...}{Additional arguments to successfully run the selected ci.appr.}
}
\value{
Returns True if passes all checks, successfully. Otherwise raises ERROR.
}
\description{
additional arguments are provided.
}
\keyword{internal}
