% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_covar_balance.R
\name{CheckCovarBalance}
\alias{CheckCovarBalance}
\title{Check covariate balance}
\usage{
CheckCovarBalance(pseudo.pop, ci.appr, ...)
}
\arguments{
\item{pseudo.pop}{The generated pseudo population. In the following format:
\itemize{
\item 1st column: outcome (Y)
\item 2nd column: exposure (w)
\item 3rd column: gps
\item 4th column to the end: covariates (c)
}}

\item{ci.appr}{The causal inference approach.}

\item{...}{Additional arguments passed to different models.}
}
\value{
Returns True if the pseudo population meet the covariate balance test
requirements.
}
\description{
Checks the covariate balance of
}
\keyword{internal}
