% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_pseudo_pop.R
\name{compile_pseudo_pop}
\alias{compile_pseudo_pop}
\title{Compile pseudo population}
\usage{
compile_pseudo_pop(
  dataset,
  ci_appr,
  gps_model = "parametric",
  bin_seq = NULL,
  nthread = 1,
  trim_quantiles,
  optimized_compile,
  ...
)
}
\arguments{
\item{dataset}{List of size 6 including the following:
\itemize{
\item Original data set + GPS values (Y, w, GPS, counter, row_index, c)
\item e_gps_pred
\item e_gps_std_pred
\item w_resid
\item gps_mx (min and max of gps)
\item w_mx (min and max of w).
}}

\item{ci_appr}{Causal inference approach.}

\item{...}{Additional parameters.}
}
\value{
\code{compile_pseudo_pop} returns the pseudo population data that is compiled based
on selected causal inference approach.
}
\description{
Compiles pseudo population based on the original population and estimated GPS
value.
}
\keyword{internal}
