% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_pseudo_pop.R
\name{CompilePseudoPop}
\alias{CompilePseudoPop}
\title{Compile pseudo population}
\usage{
CompilePseudoPop(dataset, ci.appr, ...)
}
\arguments{
\item{dataset}{List of size 6 including the following:
\itemize{
\item Original data set + GPS values (Y, w, GPS, c)
\item e_gps_pred
\item e_gps_std_pred
\item w_resid
\item gps_mx (min and max of gps)
\item w_mx (min and max of w).
}}

\item{ci.appr}{Causal inference approach.}

\item{...}{Additional parameters.}
}
\value{
\code{CompilePseudoPop} returns the pseudo population data that is compiled based
on selected causal inference approach.
}
\description{
Copiles pseudo population based on the original population and estimated GPS
value.
}
\keyword{internal}
