% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_closest_wgps.R
\name{compute_closest_wgps}
\alias{compute_closest_wgps}
\title{A function to find the closest data in subset to the original data}
\usage{
compute_closest_wgps(a, b, c, d, sc)
}
\arguments{
\item{a}{parameter description (TODO)}

\item{b}{parameter description}

\item{c}{parameter description}

\item{d}{parameter description}

\item{sc}{parameter description}
}
\value{
The function returns index of subset data that is closest to the original data
sample.
}
\description{
Function description (TODO)
}
\keyword{internal}
