% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_minmax.R
\name{compute_min_max}
\alias{compute_min_max}
\title{Compute minimum and maximum}
\usage{
compute_min_max(x)
}
\arguments{
\item{x}{vector}
}
\value{
Returns a vector of length 2. The first element is min value, and the second
element is max value.
}
\description{
Function to compute minimum and maximum of the input vector
}
\keyword{internal}
