% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_minmax.R
\name{compute_minmax}
\alias{compute_minmax}
\title{Function to compute minimum and maximum of the input vector}
\usage{
compute_minmax(x)
}
\arguments{
\item{x}{}
}
\value{
Returns a vector of length 2. The first element is min value, and the second
element is max value.
}
\description{
Function to compute minimum and maximum of the input vector
}
\keyword{internal}
