% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_outer.R
\name{compute_outer}
\alias{compute_outer}
\title{Computes distance on all possible combinations}
\usage{
compute_outer(a, b, op)
}
\arguments{
\item{a}{first vector (size n)}

\item{b}{second vector (size m)}

\item{op}{operator (e.g., '-', '+', '/', ...)}
}
\value{
A n by m matrix that includes abs difference between elements of vector a and b.
}
\description{
Computes the distance between all combination of elements in two vector. a is
vector of size n, and b is a vector of size m, the result, will be a matrix
of size(n,m)
}
\keyword{internal}
