% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_resid.R
\name{compute_resid}
\alias{compute_resid}
\title{Function to compute residual}
\usage{
compute_resid(a, b, c)
}
\arguments{
\item{a}{description}

\item{b}{description}

\item{c}{description}
}
\value{
return value (TODO)
}
\description{
Function to compute residual
}
\keyword{internal}
