% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_risk.R
\name{compute_risk}
\alias{compute_risk}
\title{Compute risk value}
\usage{
compute_risk(h, matched_Y, matched_w, w_vals)
}
\arguments{
\item{h}{A scalar representing the bandwidth value.}

\item{matched_Y}{A vector of outcome variable in the matched set.}

\item{matched_w}{A vector of continuous exposure variable in the matched set.}

\item{w_vals}{A vector of values that you want to calculate the values of
the ERF at.}
}
\value{
returns a cross-validated risk value for the input bandwidth
}
\description{
Calculates the cross-validated risk for the optimal bandwidth selection in
kernel smoothing approach.
}
\keyword{internal}
