% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_density.R
\name{ComputeDensity}
\alias{ComputeDensity}
\title{Approximate density based on another vector}
\usage{
ComputeDensity(x0, x1)
}
\arguments{
\item{x0}{vector}

\item{x1}{vector}
}
\value{
Returns approximation of density value of vector x1 based on vector x0.
}
\description{
A function to impute missing values based on density estimation of another
vector or itself after removing the missing values.
}
\keyword{internal}
