% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_risk.R
\name{ComputeRisk}
\alias{ComputeRisk}
\title{Compute risk value}
\usage{
ComputeRisk(h, matched.Y, matched.w, w.vals)
}
\arguments{
\item{h}{A scalar representing the bandwidth value.}

\item{matched.Y}{A vector of outcome variable in the matched set.}

\item{matched.w}{A vector of continuous exposure variable in the matched set.}

\item{w.vals}{A vector of values that you want to calculate the values of
the ERF at.}
}
\value{
TODO: return value
}
\description{
TODO
}
\keyword{internal}
