% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_data_into_standard_format}
\alias{convert_data_into_standard_format}
\title{Put original data into package standard data}
\usage{
convert_data_into_standard_format(Y, w, c, q1, q2, ci_appr)
}
\arguments{
\item{Y}{Output vector}

\item{w}{Treatment or exposure vector}

\item{c}{Covariate matrix}

\item{ci_appr}{Causal Inference approach}
}
\value{
Original data with place holder columns.
}
\description{
This is a temporal function to convert original data into a package standard
data. This function will be removed after addressing issue #67:
"convert accessing data from column index to column name #67"
}
\keyword{internal}
