% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weighting.R
\name{create_weighting}
\alias{create_weighting}
\title{Create pseudo population using weighting casual inference approach}
\usage{
create_weighting(dataset, ...)
}
\arguments{
\item{dataset}{The study data set.}

\item{...}{Additional parameters.}
}
\value{
Returns a data table which includes the following columns:
\itemize{
\item Y
\item w
\item gps
\item counter
\item row_index
\item ipw
\item covariates
}
}
\description{
Generates pseudo population based on weighting casual inference method.
}
\keyword{internal}
