% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matching.R
\name{CreateMatching}
\alias{CreateMatching}
\title{Create pseudo population using matching casual inference approach}
\usage{
CreateMatching(dataset, ...)
}
\arguments{
\item{dataset}{A list with 6 elements. Including An original dataset as well
as helper vectors from estimating GPS. See \code{\link[=CompilePseudoPop]{CompilePseudoPop()}} for more
details.}

\item{...}{Additional arguments passed to the function.}
}
\value{
Returns data.table of matched set.
}
\description{
Generates pseudo population based on matching casual inference method.
}
