% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_erf.R
\name{estimate_erf}
\alias{estimate_erf}
\title{Estimate smoothed exposure-response function (ERF).}
\usage{
estimate_erf(matched_Y, matched_w, bw_seq = seq(0.2, 2, 0.2), w_vals)
}
\arguments{
\item{matched_Y}{a vector of outcome variable in matched set.}

\item{matched_w}{a vector of continuous exposure variable in matched set.}

\item{bw_seq}{a vector of bandwidth values (Default is seq(0.2,2,0.2)).}

\item{w_vals}{a vector of values that you want to calculate the values of
the ERF at.}
}
\value{
\code{erf}: The function returns a vector saved the output values of
exposure-response function (ERF) given input \code{w.vals}.
}
\description{
TODO
}
\examples{

m_d <- gen_syn_data(sample_size = 100)



}
