% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gps.R
\name{estimate_gps}
\alias{estimate_gps}
\title{Estimate GPS values}
\usage{
estimate_gps(Y, w, c, pred_model, running_appr, internal_use = TRUE, ...)
}
\arguments{
\item{Y}{A vector of observed outcome variable.}

\item{w}{A vector of observed continuous exposure variable.}

\item{c}{A data frame or matrix of observed covariates variable.}

\item{pred_model}{The selected prediction model.}

\item{running_appr}{The running approach.}

\item{internal_use}{If TRUE will return helper vectors as well. Otherwise,
will return original data + GPS value.}

\item{...}{Additional arguments passed to the model.}
}
\value{
The function returns a list of 6 objects according to the following order:
\itemize{
\item Original data set + GPS values (Y, w, GPS, c)
\item e_gps_pred
\item e_gps_std_pred
\item w_resid
\item gps_mx (min and max of gps)
\item w_mx (min and max of w).
If \code{internal.use} is set to be FALSE, only originla data set + GPS will
be returend.
}
}
\description{
Estimates GPS value for each observation.
}
\examples{
m_d <- gen_syn_data(sample_size = 100)
data_with_gps <- estimate_gps(m_d$Y,
                              m_d$treat,
                              m_d[c("cf1","cf2","cf3","cf4","cf5","cf6")],
                              pred_model = "sl",
                              running_appr = "base",
                              internal_use = FALSE,
                              sl.lib = c("SL.xgboost","SL.earth","SL.gam",
                                       "SL.ranger")
                             )

}
