% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_erf.R
\name{estimate_hat_vals}
\alias{estimate_hat_vals}
\title{Estimate hat (fitted) values}
\usage{
estimate_hat_vals(bw, matched_w, w_vals)
}
\arguments{
\item{bw}{The bandwidth value.}

\item{matched_w}{A vector of continuous exposure variable in the matched set.}

\item{w_vals}{A vector of values that you want to calculate the values of the
ERF at.}
}
\value{
Returns fitted values, or the prediction made by the model for each observation.
}
\description{
Estimates the fitted values based on bandwidth value
}
\keyword{internal}
