% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hr.R
\name{estimate_hr}
\alias{estimate_hr}
\title{Estimate hazard ratio}
\usage{
estimate_hr(formula, family, data)
}
\arguments{
\item{formula}{a vector of outcome variable in matched set.}

\item{family}{a description of the error distribution (see ?gam)}

\item{data}{dataset that formula is build upon}
}
\value{
returns an object of class gnm
}
\description{
TODO
}
