% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_semi_erf.R
\name{estimate_semi_erf}
\alias{estimate_semi_erf}
\title{Estimate semi exposure-response function (semi-ERF).}
\usage{
estimate_semi_erf(formula, family, data, ci_appr)
}
\arguments{
\item{formula}{a vector of outcome variable in matched set.}

\item{family}{a description of the error distribution (see ?gam).}

\item{data}{dataset that formula is build upon.}

\item{ci_appr}{causal inference approach (matching or weighting).}
}
\value{
returns an object of class gam
}
\description{
Estimates the smoothed exposure-response function using a generalized
additive model with splines.
}
