% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gps.R
\name{EstimateGPS}
\alias{EstimateGPS}
\title{Estimate GPS values}
\usage{
EstimateGPS(y, w, c, pred.model, ...)
}
\arguments{
\item{y}{A vector of observed outcome variable (Y).}

\item{w}{A vector of observed continuous exposure variable.}

\item{c}{A data frame or matrix of observed covariates variable.}

\item{pred.model}{The selected prediction model.}

\item{...}{Additional arguments passed to the model.}
}
\value{
\code{matched_set}: The function returns a data.table saved the constructed
matched set by the proposed GPS matching approaches.
}
\description{
Estimates GPS value for each observation.
}
