% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hatvalues.R
\name{EstimateHatvals}
\alias{EstimateHatvals}
\title{Estimate hat (fitted) values}
\usage{
EstimateHatvals(bw, matched.w, w.vals)
}
\arguments{
\item{bw}{The bandwidth value.}

\item{matched.w}{A vector of continuous exposure variable in the matched set.}

\item{w.vals}{A vector of values that you want to calculate the values of the
ERF at.}
}
\value{
Returns fitted values, or the prediction made by the model for each observation.
}
\description{
Estimates the fitted values based on bandwidth value
}
\keyword{internal}
