% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_population.R
\name{gen_pseudo_pop}
\alias{gen_pseudo_pop}
\title{Generate pseudo population dataset}
\usage{
gen_pseudo_pop(
  Y,
  w,
  c,
  ci_appr,
  running_appr,
  pred_model,
  save_output = FALSE,
  save_path = NULL,
  ...
)
}
\arguments{
\item{Y}{A vector of observed outcome variable.}

\item{w}{A vector of observed continuous exposure variable.}

\item{c}{A data frame or matrix of observed covariates variable.}

\item{ci_appr}{The causal inference approach. Possible values are:
\itemize{
\item "matching": Matching by GPS
\item "weighting": Weighting by GPS
\item "adjusting": Adjusting by GPS
}}

\item{running_appr}{The running approach.}

\item{pred_model}{a prediction model}

\item{save_output}{If TRUE, output results will be stored at the save.path.
Default is FALSE.}

\item{save_path}{location for storing the final results, format of the saved
file will be detected by the file name extension.}

\item{...}{Additional arguments passed to different models.}
}
\value{
\code{GenPseudoPop} returns a data.table pseudo population that is generated
or augmented based on the selected causal inference approach (ci_appr).
}
\description{
Generates pseudo population dataset based on user defined causal inference
approach. The output dataset satisfies covariate balance requirements if
required for the selected causal inference approach.
}
\details{
\subsection{Additional parameters}{
\subsection{Causal Inference Approach (ci.appr)}{
\itemize{
\item if ci.appr = 'matching':
\itemize{
\item \emph{matching_fun}: Matching function. Available options:
\itemize{
\item matching_l1: Manhattan distance matching
}
\item \emph{delta_n}: caliper parameter.
\item \emph{scale}: a specified scale parameter to control the relative weight that
is attributed to the distance measures of the exposure versus the GPS.
\item \emph{covar_bl_method}: covariate balance method. Available options:
\itemize{
\item 'absolute'
}
\item \emph{covar_bl_trs}: covariate balance threshold
\item \emph{max_attempt}: maximum number of attempt to satisfy covariate balance.
\item See \code{\link[=create_matching]{create_matching()}} for more details about the parameters and default
values.
}
\item if ci.appr = 'weightig':
\itemize{
\item \emph{covar_bl_method}: Covariate balance method.
\item \emph{covar_bl_trs}: Covariate balance threshold
\item \emph{max_attempt}: Maximum number of attempt to satisfy covariate balance.
}
}
}

\subsection{Prediction models (pred_model)}{
\itemize{
\item if pred_model = 'sl':
\itemize{
\item \emph{sl.lib}: A vector of prediction algorithms.
}
}
}

}
}
\examples{
m_d <- gen_syn_data(sample_size = 100)
pseuodo_pop <- gen_pseudo_pop(m_d$Y,
                              m_d$treat,
                              m_d[c("cf1","cf2","cf3","cf4","cf5","cf6")],
                              ci_appr = "matching",
                              running_appr = "base",
                              pred_model = "sl",
                              sl_lib = c("SL.xgboost","SL.earth","SL.gam"),
                              covar_bl_method = "absolute",
                              covar_bl_trs = 0.1,
                              max_attempt = 1,
                              matching_fun = "matching_l1",
                              delta_n = 1,
                              scale = 0.5)

}
