% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data.R
\name{gen_syn_data}
\alias{gen_syn_data}
\title{Generate synthetic data for GPSmatching package}
\usage{
gen_syn_data(
  sample_size = 1000,
  seed = 300,
  outcome_sd = 10,
  gps_spec = 1,
  cova_spec = 1
)
}
\arguments{
\item{sample_size}{sample size description (TODO)}

\item{seed}{seed description (TODO)}

\item{outcome_sd}{outcpme_sd description (TODO)}

\item{gps_spec}{gps_spec description (TODO)}

\item{cova_spec}{cova_spec description (TODO)}
}
\value{
\code{synthetic_data}: The function returns a data.frame saved the
constructed synthetic data.
}
\description{
Generates synthetic data based on ... (TODO)
}
\examples{
s_data <- gen_syn_data(sample_size=100, seed = 403,
                                  outcome_sd = 10, gps_spec = 1,
                                  cova_spec = 1)

}
