% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsmatching_smooth.R
\name{generate_kernel}
\alias{generate_kernel}
\title{Generate Kernel function}
\usage{
generate_kernel(t)
}
\arguments{
\item{t}{standardized vector (z-score)}
}
\value{
probability distribution
}
\description{
TODO
}
\keyword{internal}
