% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data.R
\name{generate_synthetic_data}
\alias{generate_synthetic_data}
\title{Generating Synthetic Data for GPSmatching package}
\usage{
generate_synthetic_data(
  sample_size = 1000,
  seed = 300,
  outcome_sd = 10,
  gps_spec = 1,
  cova_spec = 1
)
}
\arguments{
\item{sample_size}{sample size description}

\item{seed}{seed description}

\item{outcome_sd}{outcpme_sd description}

\item{gps_spec}{gps_spec description}

\item{cova_spec}{cova_spec description}
}
\value{
\code{synthetic_data}: The function returns a data.frame saved the
 constructed synthetic data.
}
\description{
Generating Synthetic Data for GPSmatching package
}
\examples{
s_data <- generate_synthetic_data(sample_size=10000, seed = 403,
                                  outcome_sd = 10, gps_spec = 1,
                                  cova_spec = 1)

}
