% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSmatching_smooth.R
\name{GenerateKernel}
\alias{GenerateKernel}
\title{Generate Kernel function}
\usage{
GenerateKernel(t)
}
\arguments{
\item{t}{standardized vector (z-score)}
}
\value{
probability distribution
}
\description{
TODO
}
\keyword{internal}
