% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_population.R
\name{GenPseudoPop}
\alias{GenPseudoPop}
\title{Generate pseudo population dataset}
\usage{
GenPseudoPop(
  Y,
  w,
  c,
  ci.appr,
  pred.model,
  save.output = FALSE,
  save.path = NULL,
  ...
)
}
\arguments{
\item{Y}{A vector of observed outcome variable.}

\item{w}{A vector of observed continuous exposure variable.}

\item{c}{A data frame or matrix of observed covariates variable.}

\item{ci.appr}{The causal inference approach. Possible values are:
\itemize{
\item "matching": Matching by GPS
\item "weighting": Weighting by GPS
\item "adjusting": Adjusting by GPS
}}

\item{pred.model}{a prediction model}

\item{save.output}{If TRUE, output results will be stored at the save.path.
Default is FALSE.}

\item{save.path}{location for storing the final results, format of the saved
file will be detected by the file name extension.}

\item{...}{Additional arguments passed to different models.}
}
\value{
\code{GenPseudoPop} returns a data.table pseudo population that is generated
or augmented based on the selected causal inference approach (ci.appr).
}
\description{
Generates pseudo population dataset based on user defined causal inference
approach. The output dataset satisfies covariate balance requirements if
required for the selected causal inference approach.
}
\details{
\subsection{Additional parameters}{
\subsection{Causal Inference Approach (ci.appr)}{
\itemize{
\item if ci.appr = 'matching':
\itemize{
\item \emph{matching.fun}: Matching function. Available options:
\itemize{
\item matching_l1: Manhattan distance matching
}
\item \emph{delta.n}: caliper parameter.
\item \emph{scale}: a specified scale parameter to control the relative weight that
is attributed to the distance measures of the exposure versus the GPS.
\item \emph{covar.bl.method}: covariate balance method. Available options:
\itemize{
\item 'absolute'
}
\item \emph{covar.bl.trs}: covariate balance threshold
\item \emph{max.attemp}: maximum number of attempt to satisfy covariate balance.
\item See \code{\link[=CreateMatching]{CreateMatching()}} for more details about the parameters and default
values.
}
\item if ci.appr = 'weightig':
\itemize{
\item \emph{covar.bl.method}: Covariate balance method.
\item \emph{covar.bl.trs}: Covariate balance threshold
\item \emph{max.attemp}: Maximum number of attempt to satisfy covariate balance.
}
}
}

\subsection{Prediction models (pred.model)}{
\itemize{
\item if pred.model = 'sl':
\itemize{
\item \emph{sl.lib}: A vector of prediction algorithms.
}
}
}

}
}
\examples{


# Generate pseudo population with matching causal inference approach, using
# SuperLearner package, and abolute covariance balance checking.

m.d <- GenSynData(sample.size = 100)
pseuodo.pop <- GenPseudoPop(m.d$Y,
                            m.d$treat,
                            m.d[c("cf1","cf2","cf3","cf4","cf5","cf6")],
                            ci.appr = "matching",
                            pred.model = "sl",
                            sl.lib = c("SL.xgboost","SL.earth","SL.gam",
                                       "SL.ranger"),
                            covar.bl.method = "absolute",
                            covar.bl.trs = 0.1,
                            max.attemp = 1,
                            matching.fun = "MatchingL1",
                            delta.n = 1,
                            scale = 0.5)


}
