% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data.R
\name{GenSynData}
\alias{GenSynData}
\title{Generate synthetic data for GPSmatching package}
\usage{
GenSynData(
  sample.size = 1000,
  seed = 300,
  outcome.sd = 10,
  gps.spec = 1,
  cova.spec = 1
)
}
\arguments{
\item{sample.size}{sample size description (TODO)}

\item{seed}{seed description (TODO)}

\item{outcome.sd}{outcpme_sd description (TODO)}

\item{gps.spec}{gps_spec description (TODO)}

\item{cova.spec}{cova_spec description (TODO)}
}
\value{
\code{synthetic_data}: The function returns a data.frame saved the
constructed synthetic data.
}
\description{
Generates synthetic data based on ... (TODO)
}
\examples{
s_data <- GenSynData(sample.size=100, seed = 403,
                                  outcome.sd = 10, gps.spec = 1,
                                  cova.spec = 1)

}
