% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_logger.R
\name{get_logger}
\alias{get_logger}
\title{Get Logger Settings}
\usage{
get_logger()
}
\arguments{
None
}
\value{
Returns a list that includes \strong{logger_file_path} and \strong{logger_level}.
}
\description{
Returns current logger settings.
}
\examples{

set_logger("mylogger.log", "INFO")
log_meta <- get_logger()

}
