% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSmatching_smooth.R
\name{hatvals}
\alias{hatvals}
\title{TODO: Function title}
\usage{
hatvals(bw, matched_w, w.vals)
}
\arguments{
\item{bw}{bandwidth value}

\item{matched_w}{a vector of continuous exposure variable in matched set.}

\item{w.vals}{vector of values that you want to calculate the values of the ERF at.}
}
\value{
TODO: return value
}
\description{
TODO: Function title
}
\keyword{internal}
