% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSmatching_smooth.R
\name{kern_fun}
\alias{kern_fun}
\title{Kernel function}
\usage{
kern_fun(t)
}
\arguments{
\item{t}{standardized vector (z-score)}
}
\value{
probability distribution
}
\description{
Kernel function
}
\keyword{internal}
