% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_system_info}
\alias{log_system_info}
\title{Log system information}
\usage{
log_system_info()
}
\arguments{
None
}
\value{
No return value. This function is called for side effects.
}
\description{
Logs system related information into the log file.
}
\keyword{internal}
