% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSmatching_smooth.R
\name{matching_smooth}
\alias{matching_smooth}
\title{Estimate smoothed exposure-response function (ERF).}
\usage{
matching_smooth(matched_Y, matched_w, bw.seq = seq(0.2, 2, 0.2), w.vals)
}
\arguments{
\item{matched_Y}{a vector of outcome variable in matched set.}

\item{matched_w}{a vector of continuous exposure variable in matched set.}

\item{bw.seq}{a vector of bandwidth values (Default is seq(0.2,2,0.2)).}

\item{w.vals}{a vector of values that you want to calculate the values of the ERF at.}
}
\value{
\code{erf}: The function returns a vector saved the output values of exposure-response function (ERF) given input \code{w.vals}.
}
\description{
Estimate smoothed exposure-response function (ERF).
}
