% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gpsm_pspop}
\alias{plot.gpsm_pspop}
\title{Extend generic plot functions for gpsm_erf class}
\usage{
\method{plot}{gpsm_pspop}(x, ...)
}
\arguments{
\item{x}{A gpsm_erf object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\description{
A wrapper function to extend generic plot functions for gpsm_erf class.
}
