% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gpsm_erf}
\alias{print.gpsm_erf}
\title{Extend print function for gpsm_erf object}
\usage{
\method{print}{gpsm_erf}(x, ...)
}
\arguments{
\item{x}{A gpsm_erf object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Extend print function for gpsm_erf object
}
