% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gpsm_pspop}
\alias{print.gpsm_pspop}
\title{Extend print function for gpsm_pspop object}
\usage{
\method{print}{gpsm_pspop}(x, ...)
}
\arguments{
\item{x}{A gpsm_pspop object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Extend print function for gpsm_pspop object
}
