% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSmatching_smooth.R
\name{risk_fun}
\alias{risk_fun}
\title{Compute risk value}
\usage{
risk_fun(h, matched_Y, matched_w, w.vals)
}
\arguments{
\item{h}{bandwidth value}

\item{matched_Y}{a vector of outcome variable in matched set.}

\item{matched_w}{a vector of continuous exposure variable in matched set.}

\item{w.vals}{a vector of values that you want to calculate the values of the ERF at.}
}
\value{
TODO: return value
}
\description{
Compute risk value
}
\keyword{internal}
