% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_erf.R
\name{smooth_erf_kernsmooth}
\alias{smooth_erf_kernsmooth}
\title{Compute smoothed erf with kernsmooth approach}
\usage{
smooth_erf_kernsmooth(matched_Y, matched_w, matched_cw, x_eval, bw)
}
\arguments{
\item{matched_Y}{A vector of outcome value.}

\item{matched_w}{A vector of treatment value.}

\item{matched_cw}{A vector of weight or count.}

\item{bw}{A scaler number indicating the bandwidth.}
}
\value{
A vector of smoothed ERF.
}
\description{
Compute smoothed erf with kernsmooth approach
}
\keyword{internal}
