% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSmatching_smooth.R
\name{smooth_fun}
\alias{smooth_fun}
\title{Smoothing exposure response fuction based on bandwidth}
\usage{
smooth_fun(matched_Y, bw, matched_w)
}
\arguments{
\item{matched_Y}{a vector of outcome variable in matched set.}

\item{bw}{bandwidth value}

\item{matched_w}{a vector of continuous exposure variable in matched set.}
}
\value{
TODO: return value
}
\description{
Smoothing exposure response fuction based on bandwidth
}
\keyword{internal}
