% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_erf.R
\name{SmoothERF}
\alias{SmoothERF}
\title{Smooth exposure response function}
\usage{
SmoothERF(matched.Y, bw, matched.w)
}
\arguments{
\item{matched.Y}{A vector of the outcome variable in the matched set.}

\item{bw}{The bandwidth value.}

\item{matched.w}{A vector of continuous exposure variable in the matched set.}
}
\value{
Smoothed value of ERF
}
\description{
Smooths exposure response function based on bandwidth
}
\keyword{internal}
