% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.gpsm_erf}
\alias{summary.gpsm_erf}
\title{print summary of gpsm_erf object}
\usage{
\method{summary}{gpsm_erf}(object, ...)
}
\arguments{
\item{object}{A gpsm_erf object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
Returns summary of data
}
\description{
print summary of gpsm_erf object
}
