% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.gpsm_pspop}
\alias{summary.gpsm_pspop}
\title{print summary of gpsm_pspop object}
\usage{
\method{summary}{gpsm_pspop}(object, ...)
}
\arguments{
\item{object}{A gpsm_pspop object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
Returns summary of data
}
\description{
print summary of gpsm_pspop object
}
