% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_it.R
\name{train_it}
\alias{train_it}
\title{Funtction to develop prediction model based on user's preferences.}
\usage{
train_it(Y, X, model, sl.lib)
}
\arguments{
\item{Y}{Target data}

\item{X}{Input data}

\item{model}{Prediction model algorithm.
- 'sl': SuperLearner}

\item{...}{Model related parameters should be provided.}
}
\value{
prediction model
}
\description{
Funtction to develop prediction model based on user's preferences.
}
\keyword{internal}
