% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_it.R
\name{TrainIt}
\alias{TrainIt}
\title{Funtction to develop prediction model based on user's preferences.}
\usage{
TrainIt(Y, X, pred.model, ...)
}
\arguments{
\item{...}{Model related parameters should be provided.}

\item{y}{A vector of target data. TODO: change these names to avoid confusion.}

\item{x}{A vector, matrix, or dataframe of input data.}

\item{model}{Prediction model algorithm.
\itemize{
\item 'sl': SuperLearner The required parameters:
\itemize{
\item \emph{sl.lib}: a set of methods used for estimating target value (e.g.,
("SL.xgboost","SL.earth","SL.gam","SL.ranger"))
}
}}
}
\value{
prediction model
}
\description{
Funtction to develop prediction model based on user's preferences.
}
\keyword{internal}
