% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_pop.R
\name{transform_it}
\alias{transform_it}
\title{Transform data}
\usage{
transform_it(c_name, c_val, transformer)
}
\arguments{
\item{c_name}{column (attribute) name.}

\item{c_val}{column value}

\item{transformer}{transformer function.}
}
\value{
Returns transformed data.frame.
}
\description{
Transforms data into new values.
}
\keyword{internal}
