% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_gps.R
\name{trim_gps}
\alias{trim_gps}
\title{Trim a gps object based on provided trimming quantiles}
\usage{
trim_gps(gps_obj, trim_quantiles)
}
\arguments{
\item{gps_obj}{A gps S3 object.}

\item{trim_quantiles}{A vector of trimming qauntiles with two min and max
values.}
}
\value{
A trimmed gps_obj
}
\description{
Trims a gps object based on provided trimming quantiles.
}
\keyword{internal}
