% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_logger.R
\name{update_logger}
\alias{update_logger}
\title{Update logger settings}
\usage{
update_logger(logger_file_path = NULL, logger_level = NULL)
}
\arguments{
\item{logger_file_path}{A path (including file name) to log the messages.}

\item{logger_level}{The log level. Available levels include:
\itemize{
\item TRACE
\item DEBUG
\item INFO
\item SUCESS
\item WARN
\item ERROR
\item FATAL
}}
}
\description{
Updates logger settings, including log level and location of the file.
}
\examples{

update_logger("Debug")

}
