% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSmatching_smooth.R
\name{w_fun}
\alias{w_fun}
\title{TODO: Function title}
\usage{
w_fun(bw, matched_w, w.vals)
}
\arguments{
\item{bw}{bandwidth value}

\item{matched_w}{a vector of continuous exposure variable in matched set.}

\item{w.vals}{a vector of values that you want to calculate the values of the ERF at.}
}
\value{
return value (TODO)
}
\description{
TODO: Function title
}
\keyword{internal}
