% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{CheckArgsEGPS}
\alias{CheckArgsEGPS}
\title{Check EstimateGPS function arguments}
\usage{
CheckArgsEGPS(pred.model, ...)
}
\arguments{
\item{pred.model}{The selected prediction model.}

\item{...}{Addional arguments to successfully run the selected pred.model.}
}
\value{
Returns True if passes all checks, successfully. Otherwise raises ERROR.
}
\description{
Checks EstimateGPS function arguments to make sure that the required
additional arguments are provided.
}
\keyword{internal}
