% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_closest_wgps.R
\name{compute_closest_wgps}
\alias{compute_closest_wgps}
\title{Find the closest data in subset to the original data}
\usage{
compute_closest_wgps(a, b, c, d, sc)
}
\arguments{
\item{a}{Vector of the first attribute values for subset of data.}

\item{b}{Vector of the first attribute values for all data.}

\item{c}{Vector of the second attribute values for subset of data.}

\item{d}{Vector of size one for the second attribute value.}

\item{sc}{Scale parameter to give weight for two mentioned measurements.}
}
\value{
The function returns index of subset data that is closest to the original data
sample.
}
\description{
A function to compute the closest data in subset of data to the original data
based on two attributes: vector and scalar (vector of size one).
}
\keyword{internal}
