% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_resid.R
\name{compute_resid}
\alias{compute_resid}
\title{Compute Residual}
\usage{
compute_resid(a, b, c)
}
\arguments{
\item{a}{A vector}

\item{b}{A vector}

\item{c}{A vector}
}
\value{
returns a residual values.
}
\description{
Function to compute residual
}
\keyword{internal}
