% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matching.R
\name{create_matching}
\alias{create_matching}
\title{Create pseudo population using matching casual inference approach}
\usage{
create_matching(
  data_obj,
  exposure_col_name,
  bin_seq = NULL,
  gps_density = "normal",
  nthread = 1,
  ...
)
}
\arguments{
\item{data_obj}{A list of elements. Including An original dataset as well
as helper vectors from estimating GPS. See \code{\link[=compile_pseudo_pop]{compile_pseudo_pop()}} for more
details.}

\item{bin_seq}{Sequence of w (treatment) to generate pseudo population. If
NULL is passed the default value will be used, which is
\code{seq(min(w)+delta_n/2,max(w), by=delta_n)}.}

\item{gps_density}{Model type which is used for estimating GPS value, including
\code{normal} (default) and \code{kernel}.}

\item{nthread}{Number of available cores.}

\item{...}{Additional arguments passed to the function.}
}
\value{
Returns data.table of matched set.
}
\description{
Generates pseudo population based on matching casual inference method.
}
\keyword{internal}
