% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pmetric_coefs.R
\name{estimate_pmetric_coefs}
\alias{estimate_pmetric_coefs}
\title{Estimate hazard ratio}
\usage{
estimate_pmetric_coefs(formula, family, data)
}
\arguments{
\item{formula}{a vector of outcome variable in matched set.}

\item{family}{a description of the error distribution (see ?gam)}

\item{data}{dataset that formula is build upon}
}
\value{
returns an object of class gnm
}
\description{
Estimates the hazard ratios using a parametric regression model.
}
