% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gps.R
\name{EstimateGPS}
\alias{EstimateGPS}
\title{Estimate GPS values}
\usage{
EstimateGPS(Y, w, c, pred.model, internal.use = TRUE, ...)
}
\arguments{
\item{Y}{A vector of observed outcome variable.}

\item{w}{A vector of observed continuous exposure variable.}

\item{c}{A data frame or matrix of observed covariates variable.}

\item{pred.model}{The selected prediction model.}

\item{internal.use}{If TRUE will return helper vectors as well. Otherwise,
will return original data + GPS value.}

\item{...}{Additional arguments passed to the model.}
}
\value{
The function returns a list of 6 objects according to the following order:
\itemize{
\item Original data set + GPS values (Y, w, GPS, c)
\item e_gps_pred
\item e_gps_std_pred
\item w_resid
\item gps_mx (min and max of gps)
\item w_mx (min and max of w).
If \code{internal.use} is set to be FALSE, only originla data set + GPS will
be returend.
}
}
\description{
Estimates GPS value for each observation.
}
\examples{

m.d <- GenSynData(sample.size = 100)
data.with.gps <- EstimateGPS(m.d$Y,
                             m.d$treat,
                             m.d[c("cf1","cf2","cf3","cf4","cf5","cf6")],
                             pred.model = "sl",
                             internal.use = FALSE,
                             sl.lib = c("SL.xgboost","SL.earth","SL.gam",
                                       "SL.ranger")
                             )

}
