% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gps.R
\name{EstimateGPS}
\alias{EstimateGPS}
\title{Estimate GPS values}
\usage{
EstimateGPS(Y, w, c, pred.model, ...)
}
\arguments{
\item{Y}{A vector of observed outcome variable.}

\item{w}{A vector of observed continuous exposure variable.}

\item{c}{A data frame or matrix of observed covariates variable.}

\item{pred.model}{The selected prediction model.}

\item{...}{Additional arguments passed to the model.}
}
\value{
The function returns a list of 6 objects according to the following order:
\itemize{
\item Original data set + GPS values (Y, w, GPS, c)
\item e_gps_pred
\item e_gps_std_pred
\item w_resid
\item gps_mx (min and max of gps)
\item w_mx (min and max of w).
}
}
\description{
Estimates GPS value for each observation.
}
