% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalgps_smooth.R
\name{generate_kernel}
\alias{generate_kernel}
\title{Generate kernel function}
\usage{
generate_kernel(t)
}
\arguments{
\item{t}{standardized vector (z-score)}
}
\value{
probability distribution
}
\description{
Generates kernel function
}
\keyword{internal}
