% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_pop.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data}
\usage{
preprocess_data(Y, w, c, trim_quantiles, exposure_col)
}
\arguments{
\item{Y}{A data.frame of observed outcome variable:
\itemize{
\item id: Row id column.
\item Y: Outcome data.
}}

\item{w}{A data.frame of observed continuous exposure variable.
\itemize{
\item id: Row id column.
\item w: Exposure data.
}}

\item{c}{A data.frame of observed covariates variable. Also should include
id column.}

\item{trim_quantiles}{A numerical vector of two. Represents the trim quantile
level. Both numbers should be in the range of [0,1] and in increasing order
(default: c(0.01,0.99)).}

\item{exposure_col}{Column name that is used for exposure.}
}
\value{
A list with preprocessed and original data.
}
\description{
Preprocess data to isolate extra details
}
\keyword{internal}
