% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_pop.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data}
\usage{
preprocess_data(Y, w, c, trim_quantiles, exposure_col)
}
\arguments{
\item{Y}{A data.frame comprised of two columns: one contains the observed
outcome variable, and the other is labeled as 'id'. The column for the
outcome variable can be assigned any name as per your requirements.}

\item{w}{A data.frame comprised of two columns: one contains the observed
exposure variable, and the other is labeled as 'id'. The column for the
outcome variable can be assigned any name as per your requirements.}

\item{c}{A data.frame of includes observed covariate variables. It should
also consist of a column named 'id'.}

\item{exposure_col}{Column name that is used for exposure.}
}
\value{
A list with preprocessed and original data.
}
\description{
Preprocess data to isolate extra details
}
\keyword{internal}
