% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_it.R
\name{train_it}
\alias{train_it}
\title{Generate Prediction Model}
\usage{
train_it(target, input, sl_lib_internal = NULL, ...)
}
\arguments{
\item{target}{A vector of target data.}

\item{input}{A vector, matrix, or dataframe of input data.}

\item{sl_lib_internal}{The internal library to be used by SuperLearner}

\item{...}{Model related parameters should be provided.}
}
\value{
prediction model
}
\description{
Function to develop prediction model based on user's preferences.
}
\keyword{internal}
