% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_it.R
\name{train_it}
\alias{train_it}
\title{Generate Prediction Model}
\usage{
train_it(target, input, pred_model, sl_lib_internal = NULL, ...)
}
\arguments{
\item{target}{A vector of target data.}

\item{input}{A vector, matrix, or dataframe of input data.}

\item{pred_model}{Prediction model algorithm.
\itemize{
\item 'sl': SuperLearner The required parameters:
\itemize{
\item \emph{sl_lib}: a set of methods used for estimating target value (e.g.,
("SL.xgboost","SL.earth","SL.gam","SL.ranger"))
}
}}

\item{sl_lib_internal}{The internal library to be used by SuperLearner}

\item{...}{Model related parameters should be provided.}
}
\value{
prediction model
}
\description{
Function to develop prediction model based on user's preferences.
}
\keyword{internal}
